Put background style into helper function
